1.3.0

+++ Panosen.Math/Binary.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -14,7 +15,7 @@ namespace Panosen.Math
         /// <summary>
         /// 十进制转换为任意进制
         /// </summary>
-        public static List<int> ToBase(int num, int @base)
+        public static List<int> ToBase(BigInteger num, int @base)
         {
             if (num == 0)
             {
@@ -24,7 +25,7 @@ namespace Panosen.Math
             List<int> items = new List<int>();
             while (num > 0)
             {
-                items.Add(num % @base);
+                items.Add((int)(num % @base));
                 num /= @base;
             }
 
@@ -36,14 +37,14 @@ namespace Panosen.Math
         /// <summary>
         /// 任意进制转换为十进制
         /// </summary>
-        public static int FromBase(List<int> nums, int @base)
+        public static BigInteger FromBase(List<int> nums, int @base)
         {
             if (nums == null || nums.Count == 0)
             {
                 return 0;
             }
 
-            int value = 0;
+            BigInteger value = 0;
             for (int index = 0; index < nums.Count; index++)
             {
                 value += nums[index];
